backend/requirements.txt
requests==2.26.0
Django==3.2
djangorestframework==3.12.4
PyJWT==2.1.0
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
django-filter==22.1
djangorestframework-simplejwt==4.8.0

asgiref==3.3.2
gunicorn==20.0.4
psycopg2-binary==2.8.6
pytz==2020.1
sqlparse==0.3.1

python-dotenv==0.20.0
django-environ==0.8.1
django-colorfield==0.8.0
webcolors==1.13
django-measurement==3.2.4
drf-extra-fields==3.5.0
filetype==1.2.0
django-cors-headers==4.1.0
djoser==2.1.0
